refactor: make signature.rs generation-only, move request layer to protocol/request/
